fix: move functions related to Datastore to TSR-types.
This is to allow consumers of TSR-types to be able to work with TSR-types data (for example applying datastore onto a state) without having to import TSR.
