+++ packages/timeline-state-resolver/src/lib.ts
@@ -1,14 +1,5 @@
 import { klona } from 'klona'
-import {
-	Datastore,
-	Timeline,
-	TimelineDatastoreReferencesContent,
-	TSRTimelineContent,
-	ITranslatableMessage,
-	ActionExecutionResultCode,
-	TimelineDatastoreReferences,
-	ActionExecutionResult,
-} from 'timeline-state-resolver-types'
+import { ITranslatableMessage, ActionExecutionResultCode, ActionExecutionResult } from 'timeline-state-resolver-types'
 import * as _ from 'underscore'
 import { PartialDeep } from 'type-fest'
 import deepmerge = require('deepmerge')
@@ -238,46 +229,6 @@ export function endTrace(trace: Trace): FinishedTrace {
  */
 export function deferAsync(fn: () => Promise<void>, catcher: (e: unknown) => void): void {
 	fn().catch(catcher)
-}
-
-/**
- * Set a value on an object from a .-delimited path
- * @param obj The base object
- * @param path Path of the value to set
- * @param val The value to set
- */
-const set = (obj: Record<string, any>, path: string, val: any) => {
-	const p = path.split('.')
-	p.slice(0, -1).reduce((a, b) => (a[b] ? a[b] : (a[b] = {})), obj)[p.slice(-1)[0]] = val
-}
-export function fillStateFromDatastore(state: Timeline.TimelineState<TSRTimelineContent>, datastore: Datastore) {
-	// clone the state so we can freely manipulate it
-	const filledState: typeof state = JSON.parse(JSON.stringify(state))
-
-	Object.values<Timeline.ResolvedTimelineObjectInstance<TSRTimelineContent>>(filledState.layers).forEach(
-		({ content, instance }) => {
-			if ((content as TimelineDatastoreReferencesContent).$references) {
-				Object.entries<TimelineDatastoreReferences[0]>(
-					(content as TimelineDatastoreReferencesContent).$references || {}
-				).forEach(([path, ref]) => {
-					const datastoreVal = datastore[ref.datastoreKey]
-
-					if (datastoreVal !== undefined) {
-						if (ref.overwrite) {
-							// only use the datastore value if it was changed after the tl obj started
-							if ((instance.originalStart || instance.start || 0) <= datastoreVal.modified) {
-								set(content, path, datastoreVal.value)
-							}
-						} else {
-							set(content, path, datastoreVal.value)
-						}
-					}
-				})
-			}
-		}
-	)
-
-	return filledState
 }
 
 /** Create a Translatable message */
